decompose generation options
